+++ src/p_005.rs
@@ -1,5 +1,5 @@
 #[allow(dead_code)]
-pub fn problem_005() -> i128 {
+pub fn problem_005() -> i64 {
     let possible_divisors = [3, 6, 7, 8, 9, 11, 12, 13, 14, 15, 16, 17, 18, 19];
     let mut n = 40;
